Medisync/Backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pydantic==1.10.7
email-validator==2.0.0